Refactor auth integration tests to unit tests
